refactoring django project directory
