+++ pcp_alpha/Backend/pcp_app/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CapabilitiesAppConfig(AppConfig):
-    name = 'capabilities_app'
+    name = 'pcp_app'
